first run on limine bootloader

+++ src/drivers/disk/ata.c
@@ -1,7 +1,7 @@
 #include "ata.h"
 #include "../../io/io.h"
 
-// 0xE0 для Master (диск 0), 0xF0 для Slave (диск 1)
+// 0xE0
 #define ATA_MASTER 0xE0
 #define ATA_SLAVE  0xF0
 
@@ -17,9 +17,9 @@ static void ata_wait_drq() {
     while (!(inb(0x1F7) & 0x08));
 }
 
-void read_sectors_ata_pio(uint32_t target_address, uint32_t LBA, uint8_t sector_count) {
+void read_sectors_ata_pio(uint64_t target_address, uint64_t LBA, uint8_t sector_count) {
     ata_wait_bsy();
-    // ИСПОЛЬЗУЕМ 0xF0 ТАК КАК fs.img ЭТО SLAVE (index=1)
+    // slave
     outb(0x1F6, ATA_SLAVE | ((LBA >> 24) & 0x0F));
     outb(0x1F2, sector_count);
     outb(0x1F3, (uint8_t)LBA);
@@ -39,7 +39,7 @@ void read_sectors_ata_pio(uint32_t target_address, uint32_t LBA, uint8_t sector_
     }
 }
 
-void write_sectors_ata_pio(uint32_t LBA, uint8_t sector_count, uint16_t* buffer) {
+void write_sectors_ata_pio(uint64_t LBA, uint8_t sector_count, uint16_t* buffer) {
     ata_wait_bsy();
     outb(0x1F6, ATA_SLAVE | ((LBA >> 24) & 0x0F));
     outb(0x1F2, sector_count);
@@ -56,7 +56,6 @@ void write_sectors_ata_pio(uint32_t LBA, uint8_t sector_count, uint16_t* buffer)
         }
     }
     
-    // ОТПРАВЛЯЕМ КОМАНДУ СБРОСА КЭША
     outb(0x1F7, 0xE7); 
     ata_wait_bsy();
 }
