+++ src/drivers/vga/vesa.h
@@ -18,13 +18,13 @@
 #define COLOR_DARK_GREY   0x333333
 
 // Прототипы функций
-void init_vesa(uint32_t fb_addr);
-void put_pixel(int x, int y, uint32_t color);
-void draw_rect(int x, int y, int w, int h, uint32_t color);
+void init_vesa(uint64_t fb_addr);
+void put_pixel(int x, int y, uint64_t color);
+void draw_rect(int x, int y, int w, int h, uint64_t color);
 void draw_background(); // Функция для градиента
-void vesa_draw_string(const char* s, int x, int y, uint32_t fg);
-void vesa_draw_char(char c, int x, int y, uint32_t fg);
-void draw_transparent_rect(int x, int y, int w, int h, uint32_t color, uint8_t alpha);
+void vesa_draw_string(const char* s, int x, int y, uint64_t fg);
+void vesa_draw_char(char c, int x, int y, uint64_t fg);
+void draw_transparent_rect(int x, int y, int w, int h, uint64_t color, uint8_t alpha);
 void draw_equinox_logo(int start_x, int start_y);
 
 #endif
